big refactor
- restructure url.js and api.js
- modify Header.js to use url structure
- modify MainContent.js to use url structure
- correct HomPage.css style class and modify HomePage.js to use url
structure
- modify stockListPageLogic.js to use api structure
- modify stockPageLogic.js to use api structure
- modify usersPageLogic.js to use api structure
- modify stockAddFormLogic.js to use api structure and clean this file
- modify props structure in StockCard.js, add hasNavigateBtn and
hasDeleteBtn feature to this component
- add hasNavigate and hasDelete feature to StockCardList.js
- modify stockCardListLogic.js to use api structure and clean this file
- modify props structure in UserCard.js
- modify UserCardList.js to use url structure

+++ src/Components/StockCardList/StockCardList.js
@@ -3,10 +3,11 @@ import { useNavigate } from 'react-router-dom'
 import { deleteStockAsync } from './stockCardListLogic'
 import { ToastContainer } from 'react-toastify'
 import { useState, useEffect } from 'react';
+import Url from "../../url"
 
 import './StockCardList.css'
 
-const StockCardList = ({ initialStocks }) => {
+const StockCardList = ({ initialStocks, hasNavigate=true, hasDelete=true }) => {
     const [stocks, setStocks] = useState(initialStocks);
     const navigate = useNavigate()
 
@@ -14,12 +15,16 @@ const StockCardList = ({ initialStocks }) => {
         setStocks(initialStocks);
     }, [initialStocks]);
 
-    const loadViewStockPage = (stockId) => {
-        const url = `/stock/${stockId}`
+    const viewStockPage = (stockId) => {
+        if (!hasNavigate) return;
+
+        const url = Url.stock(stockId)
         navigate(url)
     }
 
     const deleteStock = async (stockId) => {
+        if (!hasDelete) return;
+        
         const success = await deleteStockAsync(stockId);
         if (success) {
             setStocks(stocks.filter(stock => stock.id !== stockId));
@@ -29,8 +34,8 @@ const StockCardList = ({ initialStocks }) => {
     return (
         <div className='stock-card-container'>
             <ToastContainer />
-            {stocks.map(stock => <StockCard key={stock.id} symbol={stock.symbol} companyName={stock.companyName} industry={stock.industry} id={stock.id}
-                onClickProcess={loadViewStockPage} deleteStockProcess={deleteStock} />)}
+            { stocks.map(stock => <StockCard key={stock.id} stockData={stock} viewStockPage={viewStockPage} 
+              deleteStockProcess={deleteStock} hasDeleteBtn={hasDelete} hasNavigateBtn={hasNavigate} />)}
         </div>
     )
 }
